requirements.txt
streamlit==1.35.0
pillow>=9.0.0